Moving timeout tests to new directory
